Working basic plot
