+++ dash/constants.py
@@ -15,6 +15,7 @@ SELECTED_COLOR_GOOD = "#008000"
 CE_COLOR = "#EF553B"
 FREE_COLOR = "#636EFA"
 GRAPHS_BACKGROUND = '#E5ECF6'
+CATEGORICAL_AXIS_ORDER = 'category descending'
 
 # LIMITS WHEN RUNNING ON WEB
 RESOURCE_LIMITED = False
